+++ src/main/java/ph/devcon/rapidpass/controllers/RegistryBatchRestController.java
@@ -1,9 +1,5 @@
 package ph.devcon.rapidpass.controllers;
 
-
-import com.opencsv.bean.ColumnPositionMappingStrategy;
-import com.opencsv.bean.CsvToBean;
-import com.opencsv.bean.CsvToBeanBuilder;
 import io.swagger.annotations.ApiParam;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,23 +7,25 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
+import ph.devcon.rapidpass.models.AgencyUser;
 import ph.devcon.rapidpass.models.RapidPassBulkData;
 import ph.devcon.rapidpass.models.RapidPassCSVdata;
+import ph.devcon.rapidpass.models.RapidPassEventLog;
 import ph.devcon.rapidpass.services.RegistryService;
+import ph.devcon.rapidpass.utilities.csv.ApproverRegistrationCsvProcessor;
+import ph.devcon.rapidpass.utilities.csv.SubjectRegistrationCsvProcessor;
 
 import javax.validation.Valid;
+import javax.validation.constraints.Min;
 import javax.validation.constraints.NotNull;
-import java.io.BufferedReader;
 import java.io.IOException;
-import java.io.InputStreamReader;
-import java.io.Reader;
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.time.ZoneOffset;
 import java.util.List;
-
 
 /**
  * Registry API Rest Controller specifically for batch operations
@@ -35,6 +33,7 @@ import java.util.List;
 @CrossOrigin
 @RestController
 @RequestMapping("/batch")
+@Validated
 @Slf4j
 public class RegistryBatchRestController {
 
@@ -49,64 +48,20 @@ public class RegistryBatchRestController {
      * Upload CSV or excel file of approved control numbers
      *
      * @param csvFile Receives CSV File Payload
+     *
      */
     @PostMapping("/access-passes")
     Iterable<String> newRequestPass(@RequestParam("file") MultipartFile csvFile)
             throws IOException, RegistryService.UpdateAccessPassException {
 
-        List<RapidPassCSVdata> approvedAccessPass;
+        SubjectRegistrationCsvProcessor processor = new SubjectRegistrationCsvProcessor();
+        List<RapidPassCSVdata> approvedAccessPass = processor.process(csvFile);
 
-        if (csvFile.isEmpty()) {
-            return null;
-        } else {
-
-            try (Reader fileReader = new BufferedReader(new InputStreamReader(csvFile.getInputStream()))) {
-                ColumnPositionMappingStrategy strategy = new ColumnPositionMappingStrategy();
-                strategy.setType(RapidPassCSVdata.class);
-                String[] accessPassCSVColumnMapping = {
-                        "passType",
-                        "aporType",
-                        "firstName",
-                        "middleName",
-                        "lastName",
-                        "suffix",
-                        "company",
-                        "idType",
-                        "identifierNumber",
-                        "plateNumber",
-                        "mobileNumber",
-                        "email",
-                        "originName",
-                        "originStreet",
-                        "originCity",
-                        "originProvince",
-                        "destName",
-                        "destStreet",
-                        "destCity",
-                        "destProvince",
-                        "remarks"
-                };
-
-                strategy.setColumnMapping(accessPassCSVColumnMapping);
-                CsvToBean<RapidPassCSVdata> csvToBean = new CsvToBeanBuilder(fileReader)
-                        .withMappingStrategy(strategy)
-                        .withType(RapidPassCSVdata.class)
-                        .withSkipLines(1)
-                        .withIgnoreLeadingWhiteSpace(true)
-                        .build();
-
-                approvedAccessPass = csvToBean.parse();
-
-                fileReader.close();
-            } catch (Exception e) {
-                throw e;
-            }
-        }
-        return this.registryService.batchUpload(approvedAccessPass);
+        return this.registryService.batchUploadRapidPassRequest(approvedAccessPass);
     }
 
     @GetMapping(value = "/access-passes", produces = {MediaType.APPLICATION_JSON_VALUE})
-    public ResponseEntity<RapidPassBulkData> downloadAccesPasses(
+    public ResponseEntity<RapidPassBulkData> downloadAccessPasses(
             @NotNull @ApiParam(value = "indicates last sync of checkpoint device in Epoch format", required = true)
             @Valid @RequestParam(value = "lastSyncOn", required = true)
                     Long lastSyncOn,
@@ -121,6 +76,30 @@ public class RegistryBatchRestController {
         Pageable pageable = PageRequest.of(pageNumber, pageSize);
 
         return ResponseEntity.ok().body(registryService.findAllApprovedSince(lastSyncDateTime, pageable));
+    }
+
+    @GetMapping("/access-pass-events")
+    public ResponseEntity<?> getAccessPassEvents(
+            @RequestParam @Min(0) Integer fromEventID,
+            @RequestParam(required = false, defaultValue = "0") Integer pageNumber,
+            @RequestParam(required = false, defaultValue = "1000") Integer pageSize)
+    {
+        Pageable page = PageRequest.of(pageNumber, pageSize);
+        RapidPassEventLog rapidPassEventLog = registryService.getAccessPassEvent(fromEventID, page);
+        if (rapidPassEventLog != null) {
+            return ResponseEntity.ok(rapidPassEventLog);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
+    }
+
+    @PostMapping("/approvers")
+    public List<String> batchRegisterApprovers(@RequestParam("file") MultipartFile csvFile) throws IOException {
+        ApproverRegistrationCsvProcessor processor = new ApproverRegistrationCsvProcessor();
+
+        List<AgencyUser> agencyUsers = processor.process(csvFile);
+
+        return this.registryService.batchUploadApprovers(agencyUsers);
     }
 }